+++ Source/CesiumRuntime/Public/CesiumFeaturesMetadataDescription.h
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 CesiumGS, Inc. and Contributors
+// Copyright 2020-2025 CesiumGS, Inc. and Contributors
 
 #pragma once